requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
assemblyai==0.36.0
attrs==24.3.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
curl_cffi==0.7.4
decorator==4.4.2
distro==1.9.0
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.5.0
g4f==0.4.1.1
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
moviepy==1.0.3
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.2.1
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.59.7
outcome==1.3.0.post0
Pillow==9.5.0
platformdirs==4.1.0
playsound==1.3.0
proglog==0.1.10
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
srt==3.5.3
srt_equalizer==0.1.8
telepot==12.7
termcolor==2.4.0
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.3.0
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
